Removed grab cursor to fix Chrome issue
Previous version included grab and grabbing hand icon if the document
cursor was set to auto. This required a bit of a hack to prevent a text
selection cursor from appearing in Framer and Safari. The hack  broke
dragging in Chrome though. So now there’s no hand cursor but things the
dragging works at least.